StepHeader 컴포넌트로 나눠서 세부구현 감추기
